+++ custom_components/ha_livekit/const.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from homeassistant.const import Platform
 
 DOMAIN = "ha_livekit"
-VERSION = "2.1.5"
+VERSION = "2.1.6"
 
 CONF_SHARED_SECRET = "shared_secret"
 CONF_PUSH_ENDPOINT_URL = "push_endpoint_url"